refactor(settings): extract reusable settings content
Move sidebar, search, and pane rendering out of the modal wrapper.

Keep dialog-only backdrop, focus trap, and close chrome in SettingsDialog.

Refs VIM-184
